Refactors main to game loop
Moves the game initialization and loop logic from `main.py` to `Game_loop.py`.

This change streamlines the game's execution flow and improves the separation of concerns.
The GUI code has been removed from main, while the game loop has been refactored into main.

This change also removes debugging code from the Game_loop file.
